+++ config/settings_local.py
@@ -18,6 +18,7 @@ EMBARGO_WRITE_GROUP = "data-topsecret-write"
 INTERNAL_READ_GROUP = "data-internal-read"
 # Minimal group for write access of internal flaws in OSIDB
 INTERNAL_WRITE_GROUP = "data-internal-write"
+
 # Contains all non-admin groups
 ALL_GROUPS = [
     *PUBLIC_READ_GROUPS,